+++ src/components/Gauge.react.js
@@ -37,7 +37,8 @@ class Gauge extends React.Component {
       id,
       className,
       style,
-      theme
+      theme,
+      digits
     } = this.props;
 
     const colorValue = getColorValue(color);
@@ -60,7 +61,7 @@ class Gauge extends React.Component {
         units={units}
         css={'transform: translateY(-150%); top: 0;'}
       >
-        {logarithmic ? log.formatValue(value, base) : value.toFixed(1)}
+        {logarithmic ? log.formatValue(value, base) : value.toFixed(digits)}
       </CurrentValue>
     );
     const filteredProps = getFilteredProps(this.props);
@@ -90,7 +91,8 @@ Gauge.defaultProps = {
   max: 10,
   base: 10,
   theme: light,
-  labelPosition: 'top'
+  labelPosition: 'top',
+  digits: 1
 };
 
 Gauge.propTypes = {
@@ -138,6 +140,11 @@ Gauge.propTypes = {
    * will be displayed
    */
   showCurrentValue: PropTypes.bool,
+
+  /**
+   * Number of digits for current value
+   */
+  digits: PropTypes.number,
 
   /**
    * Label for the current value